refactor: __init__.pyのコードを分割整理
