+++ src/streamlit_shizukani/modules/error.py
@@ -1,0 +1,5 @@
+import random
+
+
+def error():
+    return random.choice(["あんだって？", "やべえエラー出た。"])
